functions/cassandra_latest_readings/src/requirements.txt
functions-framework==3.*
cassandra-driver==3.29.2
google-cloud-storage==2.18.2
Flask==3.*
